+++ packages/react/src/components/legacy/LegacyTextField/TextField.tsx
@@ -10,11 +10,11 @@ import type {
 import { NumericFormat, PatternFormat } from 'react-number-format';
 
 import { isNumericFormat, isPatternFormat } from '../../../utilities';
-import { InputWrapper } from '../../../utilities/_InputWrapper';
+import { InputWrapper } from '../../../utilities/InputWrapper';
 import type {
   ReadOnlyVariant_,
   CharacterLimit,
-} from '../../../utilities/_InputWrapper';
+} from '../../../utilities/InputWrapper';
 
 export type LegacyTextFieldProps = {
   /**
